camera-capture/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
httpx>=0.24.0
python-dotenv>=0.19.0
